==> cuadernos/01.5_web_scraping.ipynb <==
### Este notebook recopila información sobre los días festivos en Estados Unidos entre 2019 y 2023 en el sitio web: https://www.cuandoenelmundo.com/calendario/estados-unidos/2023

from bs4 import BeautifulSoup
import requests
import pandas as pd

Creamos una función **fechar_festivos** para recopilar los días festivos en EEUU, correspondientes a los años en los que descargamos datos de vuelos, utilizando web scraping.

- Define tres listas vacías para almacenar los días, meses y los años del día festivo.
- Itera sobre los años desde 2019 hasta 2023.
- Para cada año, construye una URL correspondiente al calendario de días festivos en Estados Unidos.
- Realiza una solicitud HTTP GET a esa URL y analiza el contenido HTML utilizando BeautifulSoup.
- Busca los elementos HTML que contienen los días festivos y sus meses correspondientes, identificados por sus clases CSS.
- Extrae los días festivos y sus meses correspondientes de los elementos encontrados, y los agrega a las listas correspondientes.
- Crea un DataFrame de pandas con tres columnas: "dia", "mes" y "ano", utilizando las listas creadas anteriormente.
- Mapea los nombres de los meses a números de mes utilizando un diccionario predefinido.
- Convierte las columnas "dia", "mes" y "ano" a tipos de datos apropiados.
- Crea una nueva columna llamada "festivos" que contiene objetos ``datetime`` combinando las columnas "ano", "mes" y "dia".
- Lo guarda en un archivo pickle llamado "fecha_festivos.pkl".

def fechar_festivos():

    dias_festivos = list()
    mes_festivos = list()  
    years = list()

    for year in range(2019, 2024):

        url = f'https://www.cuandoenelmundo.com/calendario/estados-unidos/{year}'
        response = requests.get(url)
        soup = BeautifulSoup(response.text, "html.parser")

        # Dia_rojo
        reddays = soup.find_all('td', class_ = 'day redday')


        for d in reddays:
            dias_festivos.append(d.text)    

        # Mes_rojo
        redmonths = soup.find_all('td', class_ = 'month redday')


        for m in redmonths:
            mes_festivos.append(m.text)

        #Año
        for y in reddays:
            years.append(year)

    df = pd.DataFrame()
    df['dia'] = dias_festivos
    df['mes'] = mes_festivos
    df['ano'] = years

    diccionario = {'enero'     : 1,
                   'febrero'   : 2,
                   'marzo'     : 3,
                   'abril'     : 4,
                   'mayo'      : 5,
                   'junio'     : 6,
                   'julio'     : 7,
                   'agosto'    : 8,
                   'septiembre': 9,
                   'octubre'   : 10,
                   'noviembre' : 11,
                   'diciembre' : 12}
    
    df['dia'] = df['dia'].astype(int)
    df['mes'] = df['mes'].map(diccionario)
    df["festivos"] = pd.to_datetime(df["ano"].astype(str) + '-' + df["mes"].astype(str) + '-' + df["dia"].astype(str))

    df.to_pickle('fecha_festivos.pkl')

    return

fechar_festivos()
