==> cuadernos/04.3_modelo_regresion.ipynb <==
import numpy as np
import pandas as pd
import os
import matplotlib.pyplot as plt
import seaborn as sns
import plotly.express as px
from plotly.subplots import make_subplots
import plotly.graph_objects as go
from datetime import datetime, timedelta
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from collections import Counter
import plotly.express as px
from sklearn.model_selection import train_test_split

# Normalización
from sklearn.preprocessing import MinMaxScaler
from sklearn.preprocessing import StandardScaler


# Métodos de Validación
from sklearn.model_selection import train_test_split
from sklearn.model_selection import GridSearchCV

from sklearn.linear_model import LinearRegression


# Métricas para regresion
from sklearn.metrics import r2_score, mean_absolute_error, mean_squared_error


from category_encoders import TargetEncoder
from sklearn.preprocessing import LabelEncoder

df = pd.read_pickle(r'C:\Users\Nahuel\Documents\hack_a_boss\mod7-streamlit-main\proyecto_aviones\data\pickle\vuelos_limpio.pkl')

# Eliminamos las columnas que no aportan información extra y hora llegada real o retraso en la llegada para que no prediga el 100% de las veces y mes que solo tiene 1 valor
columnas_eliminar = [
    "latitude_destino",
    "latitude_origen",
    "longitude_destino",
    "longitude_origen",
    "codigo_aeropuerto_destino",
    "codigo_aeropuerto_origen",
    "direccion_destino",
    "direccion_origen",
    "retraso_llegada",
    "mes"
]
df.drop(columns=columnas_eliminar, inplace=True)


diccionario_costo = {
    'alto_costo': ["Delta Air Lines", "American Airlines", "United Airlines", "Alaska Airlines", "Hawaiian Airlines"],
    'medio_costo': ["Southwest Airlines", "JetBlue Airways"],
    'bajo_costo': ['Allegiant Air', 'Frontier Airlines', 'Spirit Airlines', 'Envoy Air', 'SkyWest Airlines', 'PSA Airlines', 
                'Endeavor Air', 'Mesa Airlines', 'Republic Airways', 'Horizon Air']}

def asignar_costo(aerolinea):
    for costo, aerolineas in diccionario_costo.items():
        if aerolinea in aerolineas:
            return costo

df['aerolinea'] = df['aerolinea'].apply(asignar_costo)

pesos_categorias = {'alto_costo': 2, 'medio_costo': 1, 'bajo_costo': 0} 

df['aerolinea'] = df['aerolinea'].map(pesos_categorias)

df["hora_salida_real"] = df["hora_salida_real"].apply(lambda x: x.hour * 60 + x.minute)
df["hora_llegada_real"] = df["hora_llegada_real"].apply(lambda x: x.hour * 60 + x.minute)
df["hora_salida_programada"] = df["hora_salida_programada"].apply(lambda x: x.hour * 60 + x.minute)

df['hora_despegue'] = df['hora_despegue'].replace('24:00', '00:00')
df['hora_despegue'] = pd.to_datetime(df['hora_despegue'], format='%H:%M').dt.hour * 60 + pd.to_datetime(df['hora_despegue'], format='%H:%M').dt.minute

# Como da error el datetime de fecha, extraemos el día del mes, que es el único que nos falta, porque mes como todos tienen el mismo no hace falta
# y sobbreescribimos la columna de fecha, 

df['fecha'] = df['fecha'].dt.day

def target_encoding(df, columna_categorica, target):

    media_columna_con_target = df.groupby(columna_categorica)[target].mean()
    df[columna_categorica] = df[columna_categorica].map(media_columna_con_target)
    return df

target_encoding(df, 'numero_cola', 'hora_llegada_real')
target_encoding(df, 'ciudad_origen', 'hora_llegada_real')
target_encoding(df, 'estado_origen', 'hora_llegada_real')
target_encoding(df, 'aeropuerto_origen', 'hora_llegada_real')
target_encoding(df, 'ciudad_destino', 'hora_llegada_real')
target_encoding(df, 'estado_destino', 'hora_llegada_real')
target_encoding(df, 'aeropuerto_destino', 'hora_llegada_real')

variables_discretas= ['aerolinea', 'fecha', 'numero_vuelo', 'numero_cola', 'hora_salida_programada', 'hora_salida_real', 'ciudad_destino', 
                      'estado_destino', 'aeropuerto_destino',  'ciudad_origen', 'estado_origen','aeropuerto_origen']
variables_continuas= ['duracion_programada_vuelo', 'duracion_real', 'distancia_millas']

# Para variables discretas (tanto numéricas como categóricas, toman un conjunto finito o numerable de valores)

def comparar_variable_discreta_con_target(dataframe, target_variable, variable_to_compare):
    fig, axes = plt.subplots(1, 2, figsize=(18, 7))
    sns.countplot(x=dataframe[variable_to_compare], ax=axes[0])
    sns.countplot(x=dataframe[variable_to_compare], hue=dataframe[target_variable], ax=axes[1])
    plt.show()

# comparar_variable_discreta_con_target(df, "hora_llegada_real", "aerolinea")

# Para variables continuas (numéricas que pueden contener valores atípicos, que pueden variar dentro de un rango específico )
def comparar_variable_continua_con_target(dataframe, target_variable, variable_to_compare):
    fig, axes = plt.subplots(2, 2, figsize=(18, 11))
    sns.boxplot(x=dataframe[variable_to_compare], ax=axes[0, 0])
    axes[0, 0].set_title(f'{variable_to_compare} Boxplot')
    sns.histplot(x=dataframe[variable_to_compare], ax=axes[0, 1])
    axes[0, 1].set_title(f'{variable_to_compare} Histogram')
    sns.boxplot(x=dataframe[target_variable], y=dataframe[variable_to_compare], ax=axes[1, 0])
    axes[1, 0].set_title(f'{variable_to_compare} vs {target_variable}')
    sns.histplot(x=dataframe[variable_to_compare], hue=dataframe[target_variable], ax=axes[1, 1])
    axes[1, 1].set_title(f'{variable_to_compare} Histogram by {target_variable}')
    plt.tight_layout()
    plt.show()

# comparar_variable_continua_con_target(df, "hora_llegada_real", "distancia_millas")

# NO CARGUEIS ESTO !!!
# os arrepentiréis xD

# comparar_variable_continua_con_target(df, "hora_llegada_real", "distancia_millas")

# VER SI QUEREMOS APLICAR CUARTILES

def calcular_cuartiles(df, variable_aplicar_cuartiles, n_inferior, n_superior):
    longitud_variable_antes = len(df[variable_aplicar_cuartiles])
    q1 = df[variable_aplicar_cuartiles].quantile(q=0.25)  # Q1
    q3 = df[variable_aplicar_cuartiles].quantile(q=0.75)  # Q3
    # Rango intercuartil (IQR)
    iqr = q3 - q1
    # Calcular los limites inferior y superior
    limite_inferior = q1 - n_inferior * iqr
    limite_superior = q3 + n_superior * iqr
    df = df[df[variable_aplicar_cuartiles].between(limite_inferior, limite_superior)]
    longitud_variable_despues = len(df[variable_aplicar_cuartiles])
    diferencia_filas = longitud_variable_antes - longitud_variable_despues
    print(f"La longitud de la variable antes era de {longitud_variable_antes}, y ahora es de {longitud_variable_despues}. Se han eliminado {diferencia_filas} filas del DF.")
    return df



# APLICO LOGARITMO

df["distancia_millas"] = np.log(df["distancia_millas"] + 1) 

# Definimos la target

TARGET = "hora_llegada_real"
X = df.drop(columns = TARGET)
y  = df[[TARGET]]

# Definimos train y test

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.20, random_state = 42, stratify = y)
print(f"X_train: {X_train.shape}, y_train: {y_train.shape}")
print(f"X_test: {X_test.shape}, y_test: {y_test.shape}")

x_scaler = StandardScaler()
X_train = x_scaler.fit_transform(X_train)
X_test = x_scaler.transform(X_test)

y_scaler = StandardScaler()
y_train = y_scaler.fit_transform(np.array(y_train).reshape(-1, 1))
y_test = y_scaler.transform(np.array(y_test).reshape(-1, 1))


Algunos algoritmos comunes utilizados en problemas de regresión incluyen la regresión lineal, regresión polinómica, regresión de vecinos más cercanos (KNN), regresión de bosques aleatorios, entre otros.

# INICIAMOS EL MODELO
from sklearn.ensemble import RandomForestRegressor

model = RandomForestRegressor()
model.fit(X_train, y_train)

yhat = model.predict(X_test)
yhat

y_test = y_scaler.inverse_transform(y_test)
yhat = y_scaler.inverse_transform(yhat.reshape(-1, 1))

def invertir_log (dato):

    return np.exp(dato) -1

df["distancia_millas"] = invertir_log(df["distancia_millas"])

r2 = r2_score(y_test, yhat)
mae = mean_absolute_error(y_test, yhat)
mse = mean_squared_error(y_test, yhat)
rmse = np.sqrt(mean_squared_error(y_test, yhat))
linear_model = [[str(model), model, r2, mae, mse, rmse]]
df_lr = pd.DataFrame(data = linear_model, columns = ["name", "model", "r2", "mae", "mse", "rmse"])
df_lr

columna = TARGET

correlaciones = df.corr()[columna]

correlaciones_significativas = correlaciones[correlaciones > 0.2]

if len(correlaciones_significativas) > 0:
    print("Las siguientes columnas tienen una correlación mayor que 0.4 con '{}' :".format(columna))
    print(correlaciones_significativas.drop(columna).index.tolist())
else:
    print("No hay columnas con una correlación mayor que 0.4 con '{}'.".format(columna))
